Move Save-Image-specific input to Save Image node
